pyproject.toml
[project]
name = "lizard-toepads"
version = "0.1.0"
description = "YOLO-based object detection for lizard toepads"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ultralytics>=8.3.0",
    "pillow>=11.0.0",
    "numpy>=2.0.0",
    "tqdm>=4.66.0",
    "opencv-python>=4.8.0",
    "pyyaml>=6.0",
    # PyTorch - install separately for GPU support
    # CUDA 12.4: uv pip install --python .venv torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
    # See README for details before running training
    "matplotlib>=3.7.0",
    "scikit-learn>=1.3.0",
    # ml-morph dependencies
    "pytorch-lightning>=2.0.0",
    "pandas>=2.0.0",
    # Note: dlib is optional, only needed for classic ml-morph workflow
    # Install manually if needed: uv pip install dlib
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = [
    "*.py",
    "README.md",
    "requirements.txt",
]
exclude = [
    "data/",
    "runs/",
    ".git/",
    "__pycache__/",
]

[tool.uv]
dev-dependencies = [
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "tensorboard>=2.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.black]
line-length = 120
target-version = ['py310']
